pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django-cors-headers>=4.7.0",
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "firebase-admin>=6.7.0",
    "requests>=2.32.3",
]
